requirements.txt
mypy==0.991
mypy-extensions==0.4.3
mysql-replication==0.31
pandas==1.5.2
psycopg2==2.9.5
PyMySQL==1.0.2
SQLAlchemy==2.0.7
typing_extensions==4.4.0
pyyaml